pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "earthquake"
version = "0.1.0"
description = "USGS earthquake medallion pipeline (Bronze/Silver/Gold)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "pandas>=2.0",
  "python-dotenv>=1.0",
  "reverse_geocoder>=1.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.4",
]

[project.scripts]
earthquake = "earthquake.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
  ".venv",
  "venv",
  "dist",
  "build",
  "__pycache__",
  ".pytest_cache",
  "data",
  "outputs",

  # Ignore notebooks (both .ipynb and Databricks-exported .py)
  "notebooks/**",
]


[tool.ruff.lint]
# Keep the linter focused but useful:
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # ignore line-length if you use a formatter

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
